+++ Changes.md
@@ -40,6 +40,19 @@ Improvements
 - PythonEditor : Added support for dropping `IECore.CompoundObject` and `IECore.ObjectMatrix` values into the code editor.
 - ShaderTweaks : Added support for creating tweaks via drag & drop from the SceneInspector to the `+` button. Dragging a single value creates an unnamed tweak of the right type. Drag a multiple selection containing the name and the value to set the value too.
 - PathListingWidget : Improved update responsiveness by updating visible items before items which are collapsed or scrolled out of view.
+- RenderManAttributes : Re-labelled displacement `Trace` checkbox as `Enabled`, to better match other DCCs.
+- USDLight :
+  - Added Arnold `lens_radius` and `aspect_ratio` parameters to the "Shaping" section.
+  - Added Arnold `shadow_density` parameter to the "Shadows" section.
+  - Renamed DiskLight `arnold:spread` plug label from "Spread" to "Spread (Arnold)".
+- Arnold :
+  - Reworked output filter handling. Now outputs with different filters can write to the same file (#6574). Also improves tidiness of resulting Arnold scene description.
+  - Allow outputs with layer names and outputs without layer names to be merged to the same file.
+- Viewer : Added `Hide`, `Unhide`, and `Prune` menu items to the right-click menu of the 3D view.
+- HierarchyView, LightEditor, AttributeEditor :
+  - Added `Hide`, `Unhide`, and `Prune` menu items.
+  - Added column for displaying and editing scene visibility.
+- LightEditor : Removed `Delete` menu item, use `Prune` instead.
 
 Fixes
 -----
@@ -48,6 +61,8 @@ Fixes
 - PathListingWidget : Fixed parent layout update when column sizes change.
 - Path : Fixed GIL management bug in `children()` binding.
 - RenderMan : Worked around RenderMan bug that prevented edits to the `render:camera` option from working during an interactive render.
+- Menu : Fixed bug causing keys pressed on the number pad to not activate keyboard shortcuts, such as the <kbd>Alt</kbd>+<kbd>[1-9]</kbd> shortcuts for assigning focus to a numeric bookmark.
+- EditScopeUI : Fixed bug causing the <kbd>H</kbd> shortcut to hide the selected locations, use <kbd>Ctrl</kbd>+<kbd>H</kbd> instead.
 
 API
 ---
@@ -58,6 +73,8 @@ API
 - AttributeEditor, HierarchyView : Added `filter` and `setFilter` plugs to the `settings()` node.
 - RenderPassEditor : Added `filter` and `hideDisabled` plugs to the `settings()` node.
 - SetEditor : Added `filter`, `hideEmptySets` and `hideEmptySelection` plugs to the `settings()` node.
+- LightEditor : Added `sceneListing()` method, to allow the context menu to be customised.
+- EditScopeAlgo : Added `setVisibility()` and `visibilityNonEditableReason()` functions.
 
 1.6.0.0 (relative to 1.5.16.2)
 =======